python/requirements.txt
paho-mqtt
pynput